Full screen implemented and camera improved
now the game lay at fullscreen but better if you do not press the Esc button, some changes to the camera now is more smooth

+++ Art_Of_Fighting/ModuleSceneking.h
@@ -44,10 +44,9 @@ public:
 	int Scene1_Fight = 0;
 	float foreground_pos;
 	bool forward;
-	int distance = 0;
+	float distance = 0;
 	iPoint positionlimitleft;//NEW
 	iPoint positionlimitright;//NEW
-
 	int font_score = -1;
 	Uint32 round_timer = 0;
